Implemented Rating of Hillfort with Stars

+++ app/src/main/java/assignment.hillfort.models/HillfortModel.kt
@@ -21,7 +21,7 @@ data class HillfortModel(@PrimaryKey(autoGenerate = true)
                          var description: String = "",
                          var visited: Boolean = false,
                          var datevisited: String = "",
-                         var rating: Long = 0,
+                         var rating: Float = 0F,
                          var favourite: Boolean = false,
                          @Embedded var location : Location = Location()) : Parcelable
